# Switch to using Pydantic `TypeAdapter`