--- a/perftest/load list of floats and ints.py
+++ b/perftest/load list of floats and ints.py
@@ -34,9 +34,8 @@ if sys.argv[1] == '--jsons':
     f = lambda: load(data, Data)
 elif sys.argv[1] == '--pydantic':
     import pydantic
-    class DataPy(pydantic.BaseModel):
-        data: List[Union[int, float]]
-    f = lambda: DataPy(**data)
+    ta = pydantic.TypeAdapter(Data)
+    f = lambda: ta.validate_python(data)
 elif sys.argv[1] == '--apischema':
     import apischema
     apischema.settings.serialization.check_type = True