Hi, I noticed a few of your comments on HN @ltworf about typedload still being faster than Pydantic V2 in a few benchmarks, so I decided to have a look at the benchmarks here.

Even without these changes, I couldn't see a benchmark that was faster with typedload. Any chance you could let me know what that benchmark was?

I updated these benchmarks to use [`TypeAdapter`](https://docs.pydantic.dev/latest/usage/models/#typeadapter) mostly to make the benchmark code cleaner. It should make a significant difference with the `dump objects.py` benchmark where we're now just doing serialization, not validation and serialization.

---

More generally, please can I gently encourage a more friendly and collaborate tone of competition?

I understand that you want to promote your library, and use comparisons to Pydantic to do that - that's fine with me. 

We're both just trying to tell people about the things we've built - let's do that in a fair and civil way.

I'm sorry about the frustration with pydantic's own benchmarks, I'm sure we (I) could have dealt with it better, but I think we ended up removing doing the right thing - removing them completely.

---

Let me know what you think, and particularly if you have any ideas about how to make Pydantic faster - would love to hear your thoughts!