@ltworf 2023-07-08
> Even without these changes, I couldn't see a benchmark that was faster with typedload. Any chance you could let me know what that benchmark was?

Have you tried running them?

It might also be that you compiled the binary specifically for your CPU model? I just downloaded whatever is on pypi.

> I updated these benchmarks to use [TypeAdapter](https://docs.pydantic.dev/latest/usage/models/#typeadapter) mostly to make the benchmark code cleaner. It should make a significant difference with the dump objects.py benchmark where we're now just doing serialization, not validation and serialization.

Since all the users of pydantic are using BaseModel, because that's what the tutorial is saying to do, I don't think it is unreasonable to benchmark BaseModel.

> More generally, please can I gently encourage a more friendly and collaborate tone of competition?

Your attitude towards denying unfavourable benchmarks for years was not friendly nor collaborative, IMHO. I never bothered to ask you to add benchmarks against typedload because I saw what happened to the others who tried; but that has set the tone for the interactions.

In any case, on a personal note, at work someone else decided to use pydantic, and since the type correctness of pydantic has never been decent (until version 2), I've had to fix hundreds of type bugs that mypy didn't catch because pydantic type annotations mean something completely different than regular python type annotations.

So you're a victim of pydantic's popularity in a way. Someone else picked it and forced me to use it, because it's popular, and I'm frustrated because I'm used to mypy being helpful rather than useless.

> We're both just trying to tell people about the things we've built - let's do that in a fair and civil way.

Have I said anything offensive or factually incorrect at any point? If so, I apologize. Could you point it out to me so that I can edit or add a comment?

> I'm sorry about the frustration with pydantic's own benchmarks, I'm sure we (I) could have dealt with it better, but I think we ended up removing doing the right thing - removing them completely.

I agree that it is not feasible for you to compare every other library, but claiming impressive speedups and offering no reproducibility doesn't seem very scientific either.

> Let me know what you think, and particularly if you have any ideas about how to make Pydantic faster - would love to hear your thoughts!

I haven't really looked into the code, so I am not familiar with how it works internally.

Please be aware that typedload is under a copyleft license, so "taking inspiration" from the code might cause copyleft to be applicable.