@ltworf 2023-07-08
For context: This is the HN post https://news.ycombinator.com/item?id=36635523