@samuelcolvin 2023-07-08
> Have you tried running them?

Yes of course.

<details>
<summary>Output on master</summary>

```
Now running: dump objects 1/10
	Running test with pydantic	0.6061511249863543 1.1641853339970112
	Running test with apischema	-1 -1
	Running test with master	0.7892320409882814 1.03329691599356
Now running: load list of ints 2/10
	Running test with pydantic	0.1505194999917876 0.22680508298799396
	Running test with apischema	-1 -1
	Running test with master	0.6379719999968074 0.6552092499914579
Now running: load list of floats and ints 3/10
	Running test with pydantic	0.40041633299551904 0.4638456669927109
	Running test with apischema	-1 -1
	Running test with master	1.6141052079910878 1.666677333007101
Now running: load list of lists 4/10
	Running test with pydantic	0.20798274999833666 0.26421087500057183
	Running test with apischema	-1 -1
	Running test with master	0.8459337500098627 0.9348077079921495
Now running: load big dictionary 5/10
	Running test with pydantic	0.49436358400271274 0.6508459580072667
	Running test with apischema	-1 -1
	Running test with master	1.5246934999886435 1.5824269589793403
Now running: load list of NamedTuple objects 6/10
	Running test with pydantic	1.2225794160040095 1.4436591250123456
	Running test with apischema	-1 -1
	Running test with master	1.4250415829883423 1.4721624169906136
Now running: load list of dataclass objects 7/10
	Running test with pydantic	0.42538149998290464 0.47644508301164024
	Running test with apischema	-1 -1
	Running test with master	2.2318695829890203 2.3449399159871973
Now running: realistic union of objects as namedtuple 8/10
	Running test with pydantic	0.902318375010509 1.0818070000095759
	Running test with apischema	-1 -1
	Running test with master	1.8282888749963604 1.8564569579903036
Now running: fail realistic union of objects as namedtuple 9/10
	Running test with pydantic	0.8789212920237333 0.9856478340225294
	Running test with apischema	-1 -1
	Running test with master	1.829223082982935 1.8403933750232682
Now running: fail load list of floats and ints 10/10
	Running test with pydantic	0.3633377499936614 0.39588808300322853
	Running test with apischema	-1 -1
	Running test with master	1.5988116249791346 1.6690446660213638
```

</details>

<details>
<summary>Output on TypAdapter branch</summary>

```
Now running: dump objects 1/10
	Running test with pydantic	0.4605952500132844 0.7565877909946721
	Running test with apischema	-1 -1
	Running test with master	0.7997676669911016 0.931802915991284
Now running: load list of ints 2/10
	Running test with pydantic	0.14689320902107283 0.18789920801646076
	Running test with apischema	-1 -1
	Running test with master	0.6314407500030939 0.6650309170072433
Now running: load list of floats and ints 3/10
	Running test with pydantic	0.41253816697280854 0.43726533302105963
	Running test with apischema	-1 -1
	Running test with master	1.5881604159949347 1.6156462090148125
Now running: load list of lists 4/10
	Running test with pydantic	0.20569249999243766 0.2572289170057047
	Running test with apischema	-1 -1
	Running test with master	0.8383092500152998 0.9299584590189625
Now running: load big dictionary 5/10
	Running test with pydantic	0.4991721250116825 0.6541437090199906
	Running test with apischema	-1 -1
	Running test with master	1.5202594159927685 1.5613216670171823
Now running: load list of NamedTuple objects 6/10
	Running test with pydantic	0.6094483750057407 0.6945809999888297
	Running test with apischema	-1 -1
	Running test with master	1.4277300829999149 1.4630528750130907
Now running: load list of dataclass objects 7/10
	Running test with pydantic	0.4293119579961058 0.48695037499419414
	Running test with apischema	-1 -1
	Running test with master	2.26074145900202 2.3371223750000354
Now running: realistic union of objects as namedtuple 8/10
	Running test with pydantic	0.5940458340046462 0.6127789999882225
	Running test with apischema	-1 -1
	Running test with master	1.8233250000048429 1.840425874979701
Now running: fail realistic union of objects as namedtuple 9/10
	Running test with pydantic	0.5761161660193466 0.5974638330226298
	Running test with apischema	-1 -1
	Running test with master	1.8142191669903696 1.82881333300611
Now running: fail load list of floats and ints 10/10
	Running test with pydantic	0.36277958398568444 0.3807309580151923
	Running test with apischema	-1 -1
	Running test with master	1.588980750006158 1.6792497909918893
```

</details>

Both typedload and pydantic were installed direct from PyPI:

```
pydantic==2.0.2
pydantic_core==2.1.2
typedload==2.24
```

I'm on MacOS M1, running Python 3.10.5. Not sure why these results differ so significantly from what you've quoted on HN?

As you can see:
* using `TypeAdapter` doesn't make a massive different in most cases
* The biggest difference is with "dump objects 1/10" - that's because the code on `master` is doing both validation and serialization in the benchmarked function for pydantic:
  * `master` Pydantic: `f = lambda: PPossessions(possessions=data.possessions).dict()`
  * `TypeAdapter` Pydantic: `f = lambda: ta.dump_python(data.possessions)`
  * typedload: `f = lambda: dump(data)`

> Since all the users of pydantic are using BaseModel, because that's what the tutorial is saying to do, I don't think it is unreasonable to benchmark BaseModel.

Of course totally up to you what you put in your benchmarks (and you're free to close this PR), but `TypedAdapter` is documented as well as `BaseModel`, `pydantic.dataclasses.dataclass` and `validate_call` - there will be more and more documentation for `TypedAdapter` as it's the most flexible way of doing validation when the type doesn't elegantly match a model.

I would say that 

```py
ta = pydantic.TypeAdapter(Data)
f = lambda: ta.validate_python(data)
```

Is by far the closest equivalent to `f = lambda: load(data, Data)`, so it would be the most useful thing for your users and potential users to see a benchmark to. 

> Your attitude towards denying unfavourable benchmarks for years was not friendly nor collaborative, IMHO.

I don't think I was unfriendly, but if I was, I'm sorry. I started off with a rule that libraries need to be more popular than pydantic for it to make sense to include them. That made sense when pydantic had 100 stars, but less sense as time went on. I'm sorry I took too long to change the rule or remove the benchmarks - everyone makes mistake.